Rename `validation` -> `validate/apitorenderer`
Move existins validators to another package.
