+++ internal/validate/apitorenderer/chart_margins.go
@@ -1,4 +1,4 @@
-package validation
+package apitorenderer
 
 import (
 	"errors"
