+++ internal/validate/apitorenderer/chart_sizes.go
@@ -1,4 +1,4 @@
-package validation
+package apitorenderer
 
 import (
 	"fmt"
